pyproject.toml
[tool.poetry]
authors = ["Stig B. Dørmænen <stigbd@gmail.com>"]
description = "A static web server for data.norge.no"
license = "Apache"
name = "static-rdf-server"
version = "0.1.0"

[tool.poetry.dependencies]
aiohttp = "^3.9.3"
aiohttp-middlewares = "^2.2.0"
certifi = "^2024.8.30"
content-negotiation = "^1.1.2"
gunicorn = "^23.0.0"
multidict = "^6.0.4"
python = ">=3.9,<3.11"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
rdflib = "^7.0.0"
pathvalidate = "^3.2.1"

[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1.1"
black = "^24.8.0"
flake8 = "^7.0.0"
flake8-annotations = "^3.0.1"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.9.16"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-import-order = "^0.18.2"
mypy = "^1.6.1"
nox = "^2023.4.22"
nox-poetry = "^1.0.3"
pep8-naming = "^0.13.3"
pyfakefs = "^5.3.0"
pytest = "^7.4.3"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
pytest-docker = "^2.0.1"
requests = "^2.31.0"
types-requests = "^2.31.0.20240125"

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict = false
warn_unreachable = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "gunicorn.*",
  "pythonjsonlogger.*",
  "content_negotiation.*",
]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.coverage.paths]
source = ["static_rdf_server", "*/site-packages"]

[tool.coverage.run]
branch = false
omit = [
  "*gunicorn_config.py",
]
source = ["static_rdf_server"]

[tool.coverage.report]
fail_under = 98
show_missing = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "unit: marks tests as unit (fast)",
  "integration: marks tests as integration (slower)",
  "contract: marks test as contract (slow)",
]
